TextureAPI/TextureAPI/requirements.txt
GitPython~=3.1.44

# Podstawowe ML
diffusers[torch]~=0.33.1
transformers~=4.52.4
accelerate~=1.7.0
safetensors~=0.5.3
peft~=0.15.2
huggingface_hub~=0.32.4
scipy~=1.15.3
#git+https://github.com/cloneofsimo/lora.git@master#egg=lora_diffusion

# GUI / GTK
PyGObject==3.44.1
pycairo==1.25.0
Pillow~=11.2.1
pillow-avif-plugin==1.4.3
scikit-image>=0.19

# API
fastapi==0.110.0
uvicorn[standard]==0.29.0
einops~=0.8.1
jsonmerge~=1.9.2
numpy~=2.2.6
requests~=2.32.3

pytest
torch~=2.7.0
gitdb~=4.0.12
typing_extensions~=4.14.0
h11~=0.16.0
filelock~=3.18.0
pip~=25.1.1
attrs~=25.3.0
wheel~=0.37.1
setuptools~=59.6.0
torchvision~=0.22.0
tqdm~=4.67.1
regex~=2024.11.6
packaging~=25.0
PyYAML~=6.0.2
idna~=3.10
zipp~=1.0.0
anyio~=4.9.0
sniffio~=1.3.1
smmap~=5.0.2
click~=8.2.1
certifi~=2025.4.26
psutil~=7.0.0
Jinja2~=3.1.6
charset-normalizer~=3.4.2
mpmath~=1.3.0
sympy~=1.14.0
triton~=3.3.0
urllib3~=2.4.0
fsspec~=2025.5.1
jsonschema~=4.24.0
referencing~=0.36.2
pydantic~=2.11.5
MarkupSafe~=2.0.1
starlette~=0.36.3
uvloop~=0.21.0
websockets~=15.0.1
imageio~=2.37.0
tifffile~=2025.5.10
lazy_loader~=0.4
networkx~=3.4.2
httptools~=0.6.4
watchfiles~=1.0.5
jsonschema-specifications~=2025.4.1
tokenizers~=0.21.1
pydantic_core~=2.33.2